api_yamdb/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
importlib-metadata==4.12.0
isort==5.10.1
Markdown==3.4.1
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
zipp==3.8.1
